todos list now propperly draw son new todo and also you can delte with a confirm prompt

+++ app/controllers/TodosController.js
@@ -55,7 +55,9 @@ export class TodosController {
 
     async deleteTodo(id) {
         try {
-            await todosService.deleteTodo(id)
+            if (window.confirm("Are you sure you want to delete this item?")) {
+                await todosService.deleteTodo(id)
+            }
         } catch (error) {
             Pop.error(error)
         }
